better route management & errors HTML page
