+++ templates/error.html
@@ -1,0 +1,28 @@
+{% extends "_layout.html" %}
+
+{% block title %}500: Server error{% endblock %}
+
+
+{% block content %}
+<div class="p-strip is-deep">
+  <div class="row u-equal-height">
+    <div class="col-6 u-vertically-center u-align--center">
+      <img src="https://assets.ubuntu.com/v1/bcdcf2c8-image-404.svg" alt="" height="365" width="360">
+    </div>
+    <div class="col-6 u-vertically-center">
+      <div>
+        <h1>{{ code }}: {{ reason }}</h1>
+        <p class="p-heading--four">Something's gone wrong.</p>
+        {% if message %}<blockquote><code>{{ message }}</code></blockquote>{% endif %}
+        {% if code != 404 %}
+          <p>
+            Try reloading the page.
+            If the error persists, please note that it may be a <a class="p-link--external" href="https://github.com/canonical-websites/assets.ubuntu.com/issues">known issue</a>.
+            If not, please <a class="p-link--external" href="https://github.com/canonical-websites/assets.ubuntu.com/issues/new">file a new issue</a>.
+          </p>
+        {% endif %}
+      </div>
+    </div>
+  </div>
+</div>
+{% endblock content %}
