+++ webapp/app.py
@@ -1,41 +1,19 @@
 # System
 import errno
-import re
-
-import flask
+import http.client
 
 # Packages
 from canonicalwebteam.flask_base.app import FlaskBase
 from flask.globals import request
 from pilbox.errors import PilboxError
 from swiftclient.exceptions import ClientException as SwiftException
+import flask
 
 # Local
 from webapp.commands import db_group, token_group
 from webapp.database import db_session
-from webapp.services import (
-    AssetAlreadyExistException,
-    AssetNotFound,
-    asset_service,
-)
-from webapp.sso import init_sso, login_required
-from webapp.views import (
-    create_asset,
-    create_redirect,
-    create_token,
-    delete_asset,
-    delete_redirect,
-    delete_token,
-    get_asset,
-    get_asset_info,
-    get_assets,
-    get_redirect,
-    get_redirects,
-    get_token,
-    get_tokens,
-    update_asset,
-    update_redirect,
-)
+from webapp.routes import api_blueprint, ui_blueprint
+from webapp.sso import init_sso
 
 app = FlaskBase(
     __name__,
@@ -53,92 +31,19 @@ app = FlaskBase(
 init_sso(app)
 
 
-@app.route("/manager/")
-@login_required
-def home():
-    query = request.values.get("q", "")
-    tag = request.values.get("tag", None)
-    asset_type = request.values.get("type")
-
-    if query or tag:
-        assets = asset_service.find_assets(
-            query=query, file_type=asset_type, tag=tag
-        )
-    else:
-        assets = []
-
-    return flask.render_template(
-        "index.html", assets=assets, query=query, type=asset_type
-    )
-
-
-@app.route("/manager/create", methods=["GET", "POST"])
-@login_required
-def create():
-    created_assets = []
-    existing_assets = []
-    failed_assets = []
-
-    if flask.request.method == "POST":
-        tags = flask.request.form.get("tags", "")
-        tags = re.split(",|\\s", tags)
-
-        optimize = flask.request.form.get("optimize", True)
-
-        for asset_file in flask.request.files.getlist("assets"):
-            try:
-                name = asset_file.filename
-                content = asset_file.read()
-
-                asset = asset_service.create_asset(
-                    file_content=content,
-                    friendly_name=name,
-                    optimize=optimize,
-                    tags=tags,
-                )
-
-                created_assets.append(asset)
-            except AssetAlreadyExistException as error:
-                asset = asset_service.find_asset(str(error))
-                if asset:
-                    existing_assets.append(asset)
-            except Exception as error:
-                failed_assets.append({"file_path": name, "error": str(error)})
-        return flask.render_template(
-            "created.html",
-            assets=created_assets,
-            existing=existing_assets,
-            failed=failed_assets,
-            tags=tags,
-            optimize=optimize,
-        )
-
-    return flask.render_template("create.html")
-
-
-@app.route("/manager/update", methods=["GET", "POST"])
-@login_required
-def update():
-    file_path = request.args.get("file-path")
-
-    if request.method == "GET":
-        asset = asset_service.find_asset(file_path)
-        if not asset:
-            flask.flash("Asset not found", "negative")
-
-    elif request.method == "POST":
-        tags = request.form.get("tags")
-        try:
-            asset = asset_service.update_asset(file_path, tags=tags.split(","))
-            flask.flash("Tags updated", "positive")
-        except AssetNotFound:
-            flask.flash("Asset not found", "negative")
-
-    return flask.render_template("update.html", asset=asset)
-
-
 # Error pages
 # ===
+def render_error(code, message):
+    # return JSON format in case of api route (with prefix /v1)
+    if request.full_path.startswith(api_blueprint.url_prefix + "/"):
+        return {"code": code, "message": message}, code
+    else:
+        return flask.render_template(
+            "error.html",
+            code=code,
+            reason=http.client.responses.get(code),
+            message=message,
+        )
 
 
 @app.errorhandler(400)
@@ -147,13 +52,13 @@ def update():
 @app.errorhandler(404)
 def error_handler(error=None):
     code = getattr(error, "code")
-    return {"status": code, "message": str(error)}, code
+    return render_error(code, str(error))
 
 
 @app.errorhandler(500)
 def error_500(error=None):
     app.extensions["sentry"].captureException()
-    return {"code": 500, "message": str(error)}, 500
+    return render_error(500, str(error))
 
 
 @app.errorhandler(OSError)
@@ -171,7 +76,7 @@ def error_os(error=None):
     if error.errno in [errno.E2BIG]:
         status = 413  # Request Entity Too Large
 
-    return {"code": status, "message": error.strerror}, status
+    return render_error(status, str(error.strerror))
 
 
 @app.errorhandler(PilboxError)
@@ -179,10 +84,7 @@ def error_pillbox(error=None):
     app.extensions["sentry"].captureException()
 
     status = error.status_code
-    return (
-        {"code": status, "message": f"Pilbox Error: {error.log_message}"},
-        500,
-    )
+    return render_error(status, f"Pilbox Error: {error.log_message}")
 
 
 @app.errorhandler(SwiftException)
@@ -195,52 +97,16 @@ def error_swift(error=None):
     elif error.msg[:12] == "Unauthorised":
         # Special case for swiftclient.exceptions.ClientException
         status = 511
+    return render_error(status, f"Swift Error: {error.msg}")
 
-    return {"code": status, "message": f"Swift Error: {error.msg}"}, status
 
-
-# API Routes
+# Apply blueprints
 # ===
-
-# Assets
-app.add_url_rule("/", view_func=get_assets)
-app.add_url_rule("/", view_func=create_asset, methods=["POST"])
-app.add_url_rule("/<path:file_path>", view_func=get_asset)
-app.add_url_rule("/<path:file_path>", view_func=update_asset, methods=["PUT"])
-app.add_url_rule(
-    "/<path:file_path>", view_func=delete_asset, methods=["DELETE"]
-)
-app.add_url_rule("/<path:file_path>/info", view_func=get_asset_info)
-
-# Tokens
-app.add_url_rule("/tokens", view_func=get_tokens)
-app.add_url_rule("/tokens/<string:name>", view_func=get_token)
-app.add_url_rule(
-    "/tokens/<string:name>", view_func=create_token, methods=["POST"]
-)
-app.add_url_rule(
-    "/tokens/<string:name>", view_func=delete_token, methods=["DELETE"]
-)
-
-# Redirects
-app.add_url_rule("/redirects", view_func=get_redirects)
-app.add_url_rule("/redirects", view_func=create_redirect, methods=["POST"])
-app.add_url_rule("/redirects/<path:redirect_path>", view_func=get_redirect)
-app.add_url_rule(
-    "/redirects/<path:redirect_path>",
-    view_func=update_redirect,
-    methods=["PUT"],
-)
-app.add_url_rule(
-    "/redirects/<path:redirect_path>",
-    view_func=delete_redirect,
-    methods=["DELETE"],
-)
+app.register_blueprint(ui_blueprint)
+app.register_blueprint(api_blueprint)
 
 # Teardown
 # ===
-
-
 @app.teardown_appcontext
 def remove_db_session(response):
     db_session.remove()
